refactor of despawnhelpers, track effect in pending state to avoid premature update loop cancellation
